Animation was occuring during a render cycle
Which it should. However, this was preventing the entity from being marked
as dirty for render.

Now we ask the object if it does animate with a check for .animates method,
which returns a bool representing if the object has or will animate in the
render method. If so, we mark the entity as dirty so that it will be rendered.