Add method to hide and do something to renderer
To Clear an entity, tell the Renderer to hide it. To clear it and
do something, tell the renderer to hide it and pass a callback. That
callback will be performed in the context of the Renderer component,
so either pass an arrow function or work from within the Renderer Component.